add authentication for invoice challenge (#45)
* copy auth template

* add authentication for invoices endpoint

* separate authentication required routes to sign/login routes

* AuthenticationForm

* Integrate Auth form to reigster and login page

* use body background

* integration with backend works

* add Summary for backend folder

* remove duplicated function

* catch duplicated email error for signing up

* replace authentication with hooks

* use query cache for invocies

* reaload page after loging out

+++ invoice/backend/src/api/invoices/invoices.service.ts
@@ -12,10 +12,10 @@ export class InvoicesService {
     private readonly helperService: HelperService,
   ) {}
 
-  getInvoices(filterStatus?: string) {
-    let filter: any = null;
+  getInvoices(userId: string, filterStatus?: string) {
+    let filter: any = { author: userId };
     if (filterStatus) {
-      filter = { status: { $in: filterStatus.split(',') } };
+      filter.status = { $in: filterStatus.split(',') };
     }
     return this.invoiceModel
       .find(filter, {
@@ -24,44 +24,55 @@ export class InvoicesService {
         clientName: 1,
         tag: 1,
         total: 1,
+        author: 1,
       })
       .then((docs) => docs.map((doc) => this.cleanDoc(doc)));
   }
 
-  getInvoice(id: string) {
-    return this.invoiceModel.findOne({ _id: id }).then((doc) => {
-      if (!doc)
-        throw new BadRequestException(`Can't find invoice with id#${id}`);
-      return this.cleanDoc(doc);
-    });
+  getInvoice(userId: string, id: string) {
+    return this.invoiceModel
+      .findOne({ _id: id, author: userId })
+      .then((doc) => {
+        if (!doc)
+          throw new BadRequestException(`Can't find invoice with id#${id}`);
+        return this.cleanDoc(doc);
+      });
   }
 
-  createInvoice(createInvoiceDto: any) {
+  createInvoice(userId: string, createInvoiceDto: any) {
     const items = createInvoiceDto.items;
     const tag = this.helperService.makeTag();
-    const data = { ...createInvoiceDto, tag, ...this.calcTotal(items) };
+    const data = {
+      ...createInvoiceDto,
+      tag,
+      ...this.calcTotal(items),
+      author: userId,
+    };
     return this.invoiceModel
       .create(data)
       .then((doc) => this.cleanDoc(doc))
       .catch((e) => e);
   }
 
-  updateInvoice(id: string, updateInvoiceDto: any) {
+  updateInvoice(userId: string, id: string, updateInvoiceDto: any) {
     let { items, ...update } = updateInvoiceDto;
     if (items) {
       update = { ...update, ...this.calcTotal(items) };
     }
     return this.invoiceModel
-      .updateOne({ _id: id }, update)
-      .then(() => this.invoiceModel.findOne({ _id: id }))
+      .findOneAndUpdate({ _id: id, author: userId }, update, {
+        returnOriginal: false, // returns the new version of doc after updating
+      })
       .then((doc) => {
         if (doc) return this.cleanDoc(doc);
-        else throw new BadRequestException(`Can't find invoice#${id}`);
+        else {
+          throw new BadRequestException(`Can't find invoice#${id} for user.`);
+        }
       });
   }
 
-  deleteInvoice(id: string) {
-    return this.invoiceModel.findOneAndDelete({ _id: id });
+  deleteInvoice(userId: string, id: string) {
+    return this.invoiceModel.findOneAndDelete({ _id: id, author: userId });
   }
 
   private cleanDoc(doc) {
